requirements.txt
# Web应用框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# 自动化测试框架
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
allure-pytest==2.13.2

# API测试
requests==2.31.0
pytest-mock==3.12.0

# 性能测试
locust==2.17.0

# 数据处理
pandas==2.1.3
openpyxl==3.1.2

# 测试数据生成
faker==19.12.0

# 数据库
# SQLite3 是Python内置模块，无需安装

# 配置管理
python-dotenv==1.0.0
PyYAML==6.0.1

# 日志和报告
coloredlogs==15.0.1
jinja2==3.1.2

# 代码质量
flake8==6.1.0
black==23.10.1

# 浏览器驱动管理
webdriver-manager==4.0.1

# 截图和图像处理
Pillow==10.1.0

# 时间处理
python-dateutil==2.8.2

# HTTP客户端
httpx==0.25.2